requirements.txt
lxml==6.0.2
PyMySQL==1.1.2
python-dotenv==1.2.2
tabulate==0.9.0
